requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
asyncio==3.4.3
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
flake8==4.0.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.8.0
pydantic==1.9.1
pyflakes==2.4.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.1
